+++ aqua/reader/reader.py
@@ -9,27 +9,23 @@ import intake
 import intake_esm
 
 import xarray as xr
-import pandas as pd
-import datetime
-from dateutil.relativedelta import relativedelta
 
 # from metpy.units import units, DimensionalityError
-import numpy as np
+
 import smmregrid as rg
 
 from aqua.util import load_yaml, load_multi_yaml
 from aqua.util import ConfigPath, area_selection
 from aqua.logger import log_configure, log_history
-from aqua.util import check_chunk_completeness, frequency_string_to_pandas
 from aqua.util import flip_lat_dir, find_vert_coord
 import aqua.gsv
 
 from .streaming import Streaming
 from .fixer import FixerMixin
 from .regrid import RegridMixin
+from .timmean import TimmeanMixin
 from .reader_utils import check_catalog_source, group_shared_dims, set_attrs
 from .reader_utils import configure_masked_fields
-
 
 # default spatial dimensions and vertical coordinates
 default_space_dims = ['i', 'j', 'x', 'y', 'lon', 'lat', 'longitude',
@@ -39,12 +35,10 @@ default_space_dims = ['i', 'j', 'x', 'y', 'lon', 'lat', 'longitude',
 # default vertical dimension
 default_vertical_dims = ['nz1', 'nz', 'level', 'height']
 
-
 # set default options for xarray
 xr.set_options(keep_attrs=True)
 
-
-class Reader(FixerMixin, RegridMixin):
+class Reader(FixerMixin, RegridMixin, TimmeanMixin):
     """General reader for NextGEMS data."""
 
     instance = None  # Used to store the latest instance of the class
@@ -148,10 +142,14 @@ class Reader(FixerMixin, RegridMixin):
 
         # get fixes dictionary and find them
         self.fix = fix  # fix activation flag
+        self.fixer_name = self.esmcat.metadata.get('fixer_name', None)
+
+        # case to disable automatic fix
+        if self.fixer_name is False:
+            self.logger.warning('A False flag is specified in fixer_name metadata, disabling fix!')
+            self.fix = False
+        
         if self.fix:
-            self.fixer_name = self.esmcat.metadata.get('fixer_name')
-            if self.fixer_name is not None:
-                self.logger.info('Fix names in metadata is %s', self.fixer_name)
             self.fixes_dictionary = load_multi_yaml(self.fixer_folder, loglevel=self.loglevel)
             self.fixes = self.find_fixes()  # find fixes for this model/exp/source
             self.dst_datamodel = datamodel
@@ -725,103 +723,6 @@ class Reader(FixerMixin, RegridMixin):
 
         return out
 
-    def timmean(self, data, freq=None, exclude_incomplete=False,
-                time_bounds=False, center_time=False):
-        """
-        Perform daily, monthly and yearly averaging.
-        Wrapper for _timmean function, which is called differently if data is a generator or not.
-        
-        Arguments:
-            data (xr.Dataset):  the input xarray.Dataset
-            freq (str):         the frequency of the time averaging.
-                                Valid values are monthly, daily, yearly. Defaults to None.
-            exclude_incomplete (bool):  Check if averages is done on complete chunks, and remove from the output
-                                        chunks which have not all the expected records.
-            time_bound (bool):  option to create the time bounds.
-            center_time (bool): option to center the time coordinate to the middle of the averaging period.
-            
-        Returns:
-            A xarray.Dataset containing the time averaged data.
-        """
-        if isinstance(data, types.GeneratorType):
-            return self._timmeangen(data, freq=freq, exclude_incomplete=exclude_incomplete,
-                                    time_bounds=time_bounds, center_time=center_time)
-        else:
-            return self._timmean(data, freq=freq, exclude_incomplete=exclude_incomplete,
-                                 time_bounds=time_bounds, center_time=center_time)
-
-    def _timmeangen(self, data, **kwargs):
-        """Call the timmean function for iterators"""
-        for ds in data:
-            yield self._timmean(ds, **kwargs)
-
-    def _timmean(self, data, freq=None, exclude_incomplete=False,
-                 time_bounds=False, center_time=False):
-        """
-        Perform daily, monthly and yearly averaging.
-        See timmean for more details.
-        """
-        resample_freq = frequency_string_to_pandas(freq)
-
-        # Get original frequency (for history)
-        if len(data.time) > 1:
-            orig_freq = data['time'].values[1]-data['time'].values[0]
-            # Convert time difference to hours
-            self.orig_freq = np.timedelta64(orig_freq, 'ns') / np.timedelta64(1, 'h')
-        else:
-            self.logger.warning('A single timestep is available, is this correct?')
-            self.orig_freq = 'Unknown'
-            if exclude_incomplete:
-                self.logger.warning('Disabling exclude incomplete since it cannot work if we have a single tstep!')
-                exclude_incomplete = False
-
-        try:
-            # Resample to the desired frequency
-            self.logger.info('Resampling to %s frequency...', str(resample_freq))
-            out = data.resample(time=resample_freq).mean()
-        except ValueError as exc:
-            raise ValueError('Cant find a frequency to resample, aborting!') from exc
-
-        # Set time:
-        # if not center_time as the first timestamp of each month/day according to the sampling frequency
-        # if center_time as the middle timestamp of each month/day according to the sampling frequency
-        # TODO: extend it to other frequencies
-        if center_time:
-            if resample_freq == 'Y' or resample_freq == '1Y':
-                self.logger.debug("Setting time to the middle of the year")
-                offset = pd.DateOffset(months=6) # TODO: expand it to other frequencies
-                out['time'] = out['time'].to_index().to_period(resample_freq).to_timestamp() + offset
-            else:
-                self.logger.error("center_time is not implemented yet for frequency %s", resample_freq)
-                out['time'] = out['time'].to_index().to_period(resample_freq).to_timestamp().values
-        else:
-            self.logger.debug("Setting time to the beginning of the year")
-            out['time'] = out['time'].to_index().to_period(resample_freq).to_timestamp().values
-
-        if exclude_incomplete:
-            boolean_mask = check_chunk_completeness(data, resample_frequency=resample_freq)
-            out = out.where(boolean_mask, drop=True)
-
-        # Check time is correct
-        if np.any(np.isnat(out.time)):
-            raise ValueError('Resampling cannot produce output for all frequency step, is your input data correct?')
-
-        out = log_history(out, f"resampled from frequency {self.orig_freq} h to frequency {resample_freq} by AQUA timmean")
-
-        # Add a variable to create time_bounds
-        if time_bounds:
-            resampled = data.time.resample(time=resample_freq)
-            time_bnds = xr.concat([resampled.min(), resampled.max()], dim='bnds').transpose()
-            time_bnds['time'] = out.time
-            time_bnds.name = 'time_bnds'
-            out = xr.merge([out, time_bnds])
-            if np.any(np.isnat(out.time_bnds)):
-                raise ValueError('Resampling cannot produce output for all time_bnds step!')
-            log_history(out, "time_bnds added by by AQUA timmean")
-
-        out.aqua.set_default(self)  # This links the dataset accessor to this instance of the Reader class
-
-        return out
 
     def _check_if_regridded(self, data):
         """
@@ -1154,8 +1055,8 @@ class Reader(FixerMixin, RegridMixin):
 
         if self.fix:
             print("Data fixing is active:")
-            if "fix_family" in metadata.keys():
-                print("  Fix family is %s" % metadata["fix_family"])
+            if "fixer_name" in metadata.keys():
+                print("  Fixer name is %s" % metadata["fixer_name"])
             else:
                 # TODO: to be removed when all the catalogues are updated
                 print("  Fixes: %s" % self.fixes)