+++ CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 Unreleased in the current development version.
 
+- `retrieve_plain()` method now set off startdate and enddate (#829)
+- Complete restructure of fixer to make use of `fixer_name`: set a default and a `False` (#746)
+- Added `center_time` option in the `timmean()` method to save the time coordinate in the middle of the time interval and create a Timmean module and related TimmeanMixin class (#811)
+- Fixer to rename coordinates available (#822)
+- Fixing new pandas timedelta: replacing H with h in all FDB catalog (#786)
+- Adding a run test label to trigger CI (#826)
+
 ## [v0.6.2]
 
 Complete list:
@@ -16,7 +23,7 @@ Complete list:
 - LRA to use `auto` final time and `exclude_incomplete` (#791)
 - Hotfix for v0.12.0 of the GSV_interface related to valid_time (#788)
 - Global time series adapted to new data governance (#785)
-- AtmoGlobalMean diagnostic improvements and adaptation to new data governance (#745 #789 #807)
+- AtmoGlobalMean diagnostic improvements and adaptation to new data governance (#745 #789 #807 #812)
 - Sea-ice diagnostic adapted to new data governance (#790)
 - Implement a fix setting to NaN the data of the first step in each month (for IFS historical-1990) (#776)